.delay() bugfix and extension first try
It needs more testing, but works on the added tests and also is backwards compatible.


Updated description .delay()

.delay( [ delayId ], duration, [ queueName ] )
delayId = An optional string containing the name of the delay id. The delay can be interrupted by using .removeDelay( delayId ).
duration = An integer indicating the number of milliseconds to delay execution of the next item in the queue.
queueName = A string containing the name of the queue. Defaults to fx, the standard effects queue.
Updated description .clearQueue()

.clearQueue( [ queueName ] )
queueNameA = string containing the name of the queue. Defaults to fx, the standard effects queue.

Also clears all delays of the element.
New description .removeDelay( [ delayId ], [ queueName ] )

.delay( [ delayId ], duration, [ queueName ] )
delayId = An optional string containing the name of the delay id. If no delayId is given, the last delay of the element will be canceled.
queueName = A string containing the name of the queue. Defaults to fx, the standard effects queue.

Currently you need to clear all delays of an element, before you can append new operations the the element or you call .clearQueue() first.



There are currently two things that need improvement:

1. I tried to generate a unique id for all timeouts for an element, if no timeout id is given. This is currently done with: 

(type + jQuery.now() + Math.random() * 9999999 )

But it's very lame, maybe sometimes random will generate the same number. Is there any real unique id generator implemented alredy in jquery?

2. If removeDelay(delayId) is called, removeQueue() is called, if the last delay of the element is removed. This is also very lame. I need to add a functionality that splices a position out of the queue. Currently the queue needs to be cleaned, before new jquery operations can be applied to an element, because by removing the timeout within a queue, the whole queue will become invalid.